requirements.txt
openai
langchain
jupyter
python-dotenv
langchain-community
langchain-openai
nemoguardrails
pandas